Move junit rules
