+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/MySqlJdbcDistributedTest.java
@@ -25,8 +25,8 @@ import java.sql.Types;
 import org.junit.ClassRule;
 
 import org.apache.geode.test.dunit.rules.MemberVM;
-import org.apache.geode.test.junit.rules.MySqlConnectionRule;
-import org.apache.geode.test.junit.rules.SqlDatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.MySqlConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.SqlDatabaseConnectionRule;
 
 /**
  * End-to-end dunits for jdbc connector
