+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/MySqlJdbcWriterIntegrationTest.java
@@ -20,8 +20,8 @@ import java.sql.SQLException;
 
 import org.junit.ClassRule;
 
-import org.apache.geode.test.junit.rules.DatabaseConnectionRule;
-import org.apache.geode.test.junit.rules.MySqlConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.DatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.MySqlConnectionRule;
 
 public class MySqlJdbcWriterIntegrationTest extends JdbcWriterIntegrationTest {
