+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/PostgresJdbcLoaderIntegrationTest.java
@@ -31,8 +31,8 @@ import org.apache.geode.cache.Region;
 import org.apache.geode.connectors.jdbc.internal.configuration.FieldMapping;
 import org.apache.geode.pdx.FieldType;
 import org.apache.geode.pdx.PdxInstance;
-import org.apache.geode.test.junit.rules.DatabaseConnectionRule;
-import org.apache.geode.test.junit.rules.PostgresConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.DatabaseConnectionRule;
+import org.apache.geode.connectors.jdbc.rules.PostgresConnectionRule;
 
 public class PostgresJdbcLoaderIntegrationTest extends JdbcLoaderIntegrationTest {
