--- geode-connectors/src/main/resources/org/apache/geode/internal/sanctioned-geode-connectors-serializables.txt
@@ -1,8 +1,0 @@
-org/apache/geode/connectors/jdbc/JdbcConnectorException,true,1
-org/apache/geode/connectors/jdbc/internal/RegionMappingExistsException,false
-org/apache/geode/connectors/jdbc/internal/RegionMappingNotFoundException,false
-org/apache/geode/connectors/jdbc/internal/cli/CreateMappingFunction,false
-org/apache/geode/connectors/jdbc/internal/cli/CreateMappingPreconditionCheckFunction,false
-org/apache/geode/connectors/jdbc/internal/cli/DestroyMappingFunction,false
-org/apache/geode/connectors/jdbc/internal/configuration/FieldMapping,false,jdbcName:java/lang/String,jdbcNullable:boolean,jdbcType:java/lang/String,pdxName:java/lang/String,pdxType:java/lang/String
-org/apache/geode/connectors/jdbc/internal/configuration/RegionMapping,false,catalog:java/lang/String,dataSourceName:java/lang/String,fieldMappings:java/util/List,ids:java/lang/String,pdxName:java/lang/String,regionName:java/lang/String,schema:java/lang/String,specifiedIds:boolean,tableName:java/lang/String
